# rrule iterator speedup