--- a/util.go
+++ b/util.go
@@ -66,7 +66,7 @@ func contains(list []int, elem int) bool {
 
 func timeContains(list []time.Time, elem time.Time) bool {
 	for _, t := range list {
-		if t == elem {
+		if t.Equal(elem) {
 			return true
 		}
 	}
@@ -176,4 +176,9 @@ func after(next Next, dt time.Time, inc bool) time.Time {
 			return v
 		}
 	}
+}
+
+type optInt struct {
+	Int     int
+	Defined bool
 }